Pass type hints (mypy check)
