+++ parallax/cameras/camera_base_binding.py
@@ -135,10 +135,11 @@ class BaseCamera(ABC):
     Abstract base class for camera operations.
     Defines the interface expected from all camera types.
     """
+    last_capture_time: float
 
-    def __init__(self):
+    def __init__(self) -> None:
         # Every camera implementation must assign an instance of BaseSettings here
-        self.settings: BaseSettings = None
+        self.settings: BaseSettings | None = None
 
     @abstractmethod
     def name(self, sn_only: bool = False) -> str:
