+++ parallax/config/config_calibration.py
@@ -17,8 +17,8 @@ Y_COORDS = Y_COORDS_HALF * 2 + 1
 OBJPOINTS = np.zeros((X_COORDS + Y_COORDS, 3), np.float32)
 OBJPOINTS[:X_COORDS, 0] = np.arange(-X_COORDS_HALF, X_COORDS_HALF + 1)
 OBJPOINTS[X_COORDS:, 1] = np.arange(-Y_COORDS_HALF, Y_COORDS_HALF + 1)
-OBJPOINTS = OBJPOINTS * WORLD_SCALE
-OBJPOINTS = np.around(OBJPOINTS, decimals=2)
+OBJPOINTS *= WORLD_SCALE
+np.around(OBJPOINTS, decimals=2, out=OBJPOINTS)
 
 CENTER_INDEX_X = X_COORDS_HALF
 CENTER_INDEX_Y = X_COORDS + Y_COORDS_HALF
